move a redcarpet render file into a namespace-appropriate directory
