+++ lib/cdo/redcarpet/render/inline.rb
